# openapi.yaml
openapi: 3.1.0
info:
  title: ViAna Orchestrator API
  version: 0.1.0
  description: |
    Job management API for ViAna moving-count engine.
    Machine-readable contracts: `packages/contracts/schemas/`.
    Human reference: `docs/api_contracts.md`.
servers:
  - url: http://localhost:8000
    description: Local Docker publish (default)

paths:
  /:
    get:
      summary: Service metadata
      operationId: getRoot
      responses:
        "200":
          description: Orchestrator identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceInfo"

  /health:
    get:
      summary: Health check
      operationId: getHealth
      tags: [health]
      responses:
        "200":
          description: Service healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /utils/prescan:
    post:
      summary: Prescan video (OCR + line proposal)
      operationId: postPrescan
      tags: [utils]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescanRequest"
      responses:
        "200":
          description: Prescan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescanResponse"
        "501":
          description: Unused (Phase 6 implemented)

  /utils/prescan/{prescan_id}/preview.jpg:
    get:
      summary: Prescan preview JPEG
      operationId: getPrescanPreview
      tags: [utils]
      parameters:
        - name: prescan_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: JPEG preview
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Preview not found
    get:
      summary: List jobs
      operationId: listJobs
      tags: [jobs]
      parameters:
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            pattern: "^[a-z0-9][a-z0-9_-]*$"
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobStatus"
        "501":
          description: Not implemented (Phase 6; blocked on Phase 5 engine)
    post:
      summary: Submit moving-count job
      operationId: postJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobSubmitRequest"
      responses:
        "201":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobSubmitResponse"
        "409":
          description: Checkpoint exists; explicit resume required
        "422":
          description: Invalid body (including client-supplied job_id or gpu_device)
        "501":
          description: Not implemented (Phase 6; blocked on Phase 5 engine)

  /jobs/{job_id}:
    get:
      summary: Get job status
      operationId: getJob
      tags: [jobs]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
        "404":
          description: Job not found
        "501":
          description: Not implemented (Phase 6; blocked on Phase 5 engine)
    delete:
      summary: Cancel job worker
      operationId: deleteJob
      tags: [jobs]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Cancelled
        "501":
          description: Not implemented (Phase 6; blocked on Phase 5 engine)

  /jobs/{job_id}/resume:
    post:
      summary: Explicit resume from checkpoint
      operationId: resumeJob
      tags: [jobs]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Job resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobSubmitResponse"
        "501":
          description: Not implemented (Phase 6; blocked on Phase 5 engine)

  /jobs/{job_id}/start-fresh:
    post:
      summary: Delete checkpoint and restart
      operationId: startFreshJob
      tags: [jobs]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Job restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobSubmitResponse"
        "501":
          description: Not implemented (Phase 6; blocked on Phase 5 engine)

  /jobs/{job_id}/aggregate:
    post:
      summary: Rebuild 15-minute aggregation CSV
      operationId: aggregateJob
      tags: [jobs]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Aggregation rebuilt
        "501":
          description: Not implemented (blocked on Phase 2 viana aggregate)

  /ws/jobs:
    get:
      summary: Job telemetry WebSocket
      operationId: jobsTelemetry
      tags: [jobs]
      description: |
        WebSocket. Payload matches packages/contracts/schemas/telemetry.schema.json.
        Stub until Phase 5 engine stdout exists.
      responses:
        "101":
          description: Switching protocols
        "501":
          description: Not implemented (Phase 6; blocked on Phase 5 engine)

  /projects/{project_id}/profiles:
    get:
      summary: List calibration profiles
      operationId: listProfiles
      tags: [profiles]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9][a-z0-9_-]*$"
      responses:
        "200":
          description: Profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalibrationProfile"
        "501":
          description: Not implemented (Phase 6)
    post:
      summary: Create calibration profile
      operationId: createProfile
      tags: [profiles]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9][a-z0-9_-]*$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalibrationProfile"
      responses:
        "201":
          description: Profile stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalibrationProfile"
        "501":
          description: Not implemented (Phase 6)

components:
  schemas:
    ServiceInfo:
      type: object
      required: [service, version, phase]
      properties:
        service:
          type: string
          example: viana-orchestrator
        version:
          type: string
          example: "0.1.0"
        phase:
          type: integer
          example: 6

    HealthResponse:
      type: object
      required: [status, phase]
      properties:
        status:
          type: string
          enum: [ok]
        phase:
          type: integer

    PrescanRequest:
      type: object
      required: [source_video_path, project_id]
      properties:
        source_video_path:
          type: string
        project_id:
          type: string
          pattern: "^[a-z0-9][a-z0-9_-]*$"
        frame_offset_sec:
          type: number
          default: 0

    PrescanResponse:
      type: object
      description: See packages/contracts/schemas/prescan_response.schema.json

    JobSubmitRequest:
      type: object
      description: See packages/contracts/schemas/job_submit.schema.json

    JobSubmitResponse:
      type: object
      required: [job_id, status, gpu_device, queue_position, output_dir]
      properties:
        job_id:
          type: string
        status:
          $ref: "#/components/schemas/JobStatusEnum"
        gpu_device:
          type: string
          pattern: "^cuda:[01]$"
        queue_position:
          type: integer
          minimum: 0
        output_dir:
          type: string

    JobStatus:
      type: object
      description: See packages/contracts/schemas/job_status.schema.json

    JobStatusEnum:
      type: string
      enum: [PENDING, PROCESSING, PAUSED, COMPLETED, FAILED, CANCELLED]

    CalibrationProfile:
      type: object
      description: See packages/contracts/schemas/calibration_profile.schema.json

    TelemetryMessage:
      type: object
      description: See packages/contracts/schemas/telemetry.schema.json